Reverted extraction of buildQuery
